move large fixture to its own file
