+++ __tests__/serverjs/cubefn.test.js
@@ -2,167 +2,10 @@ var sinon = require("sinon");
 const cubefn = require("../../serverjs/cubefn");
 const carddb = require("../../serverjs/cards");
 const cubefixture = require("../../fixtures/examplecube");
+const landsfixture = require("../../fixtures/examplelands");
 let Cube = require('../../models/cube');
 
 const fixturesPath = "fixtures";
-
-const exampleBasics = {
-  "plains": {
-    color_identity: ['W'],
-    set: 'unh',
-    collector_number: '136',
-    promo: false,
-    digital: false,
-    border_color: 'black',
-    name: 'Plains',
-    name_lower: 'plains',
-    full_name: 'Plains [unh-136]',
-    artist: 'John Avon',
-    scryfall_uri: 'https://scryfall.com/card/unh/136/plains?utm_source=api',
-    rarity: 'common',
-    oracle_text: '({T}: Add {W}.)',
-    _id: '1d7dba1c-a702-43c0-8fca-e47bbad4a00f',
-    cmc: 0,
-    legalities: {
-      Legacy: true,
-      Modern: true,
-      Standard: true,
-      Pauper: true
-    },
-    parsed_cost: [''],
-    colors: [],
-    type: 'Basic Land — Plains',
-    tcgplayer_id: 37909,
-    image_small: 'https://img.scryfall.com/cards/small/front/1/d/1d7dba1c-a702-43c0-8fca-e47bbad4a00f.jpg?1565989378',
-    image_normal: 'https://img.scryfall.com/cards/normal/front/1/d/1d7dba1c-a702-43c0-8fca-e47bbad4a00f.jpg?1565989378',
-    art_crop: 'https://img.scryfall.com/cards/art_crop/front/1/d/1d7dba1c-a702-43c0-8fca-e47bbad4a00f.jpg?1565989378',
-    colorcategory: 'l',
-  },
-  "mountain": {
-    color_identity: ['R'],
-    set: 'unh',
-    collector_number: '139',
-    promo: false,
-    digital: false,
-    border_color: 'black',
-    name: 'Mountain',
-    name_lower: 'mountain',
-    full_name: 'Mountain [unh-139]',
-    artist: 'John Avon',
-    scryfall_uri: 'https://scryfall.com/card/unh/139/mountain?utm_source=api',
-    rarity: 'common',
-    oracle_text: '({T}: Add {R}.)',
-    _id: '42232ea6-e31d-46a6-9f94-b2ad2416d79b',
-    cmc: 0,
-    legalities: {
-      Legacy: true,
-      Modern: true,
-      Standard: true,
-      Pauper: true
-    },
-    parsed_cost: [''],
-    colors: [],
-    type: 'Basic Land — Mountain',
-    tcgplayer_id: 37896,
-    image_small: 'https://img.scryfall.com/cards/small/front/4/2/42232ea6-e31d-46a6-9f94-b2ad2416d79b.jpg?1565989372',
-    image_normal: 'https://img.scryfall.com/cards/normal/front/4/2/42232ea6-e31d-46a6-9f94-b2ad2416d79b.jpg?1565989372',
-    art_crop: 'https://img.scryfall.com/cards/art_crop/front/4/2/42232ea6-e31d-46a6-9f94-b2ad2416d79b.jpg?1565989372',
-    colorcategory: 'l',
-  },
-  "forest": {
-    color_identity: ['G'],
-    set: 'unh',
-    collector_number: '140',
-    promo: false,
-    digital: false,
-    border_color: 'black',
-    name: 'Forest',
-    name_lower: 'forest',
-    full_name: 'Forest [unh-140]',
-    artist: 'John Avon',
-    scryfall_uri: 'https://scryfall.com/card/unh/140/forest?utm_source=api',
-    rarity: 'common',
-    oracle_text: '({T}: Add {G}.)',
-    _id: '19e71532-3f79-4fec-974f-b0e85c7fe701',
-    cmc: 0,
-    legalities: {
-      Legacy: true,
-      Modern: true,
-      Standard: true,
-      Pauper: true
-    },
-    parsed_cost: [''],
-    colors: [],
-    type: 'Basic Land — Forest',
-    tcgplayer_id: 37859,
-    image_small: 'https://img.scryfall.com/cards/small/front/1/9/19e71532-3f79-4fec-974f-b0e85c7fe701.jpg?1565989358',
-    image_normal: 'https://img.scryfall.com/cards/normal/front/1/9/19e71532-3f79-4fec-974f-b0e85c7fe701.jpg?1565989358',
-    art_crop: 'https://img.scryfall.com/cards/art_crop/front/1/9/19e71532-3f79-4fec-974f-b0e85c7fe701.jpg?1565989358',
-    colorcategory: 'l',
-  },
-  "swamp": {
-    color_identity: ['B'],
-    set: 'unh',
-    collector_number: '138',
-    promo: false,
-    digital: false,
-    border_color: 'black',
-    name: 'Swamp',
-    name_lower: 'swamp',
-    full_name: 'Swamp [unh-138]',
-    artist: 'John Avon',
-    scryfall_uri: 'https://scryfall.com/card/unh/138/swamp?utm_source=api',
-    rarity: 'common',
-    oracle_text: '({T}: Add {B}.)',
-    _id: '8365ab45-6d78-47ad-a6ed-282069b0fabc',
-    cmc: 0,
-    legalities: {
-      Legacy: true,
-      Modern: true,
-      Standard: true,
-      Pauper: true
-    },
-    parsed_cost: [''],
-    colors: [],
-    type: 'Basic Land — Swamp',
-    tcgplayer_id: 37935,
-    image_small: 'https://img.scryfall.com/cards/small/front/8/3/8365ab45-6d78-47ad-a6ed-282069b0fabc.jpg?1565989387',
-    image_normal: 'https://img.scryfall.com/cards/normal/front/8/3/8365ab45-6d78-47ad-a6ed-282069b0fabc.jpg?1565989387',
-    art_crop: 'https://img.scryfall.com/cards/art_crop/front/8/3/8365ab45-6d78-47ad-a6ed-282069b0fabc.jpg?1565989387',
-    colorcategory: 'l',
-  },
-  "island": {
-    color_identity: ['U'],
-    set: 'unh',
-    collector_number: '137',
-    promo: false,
-    digital: false,
-    border_color: 'black',
-    name: 'Island',
-    name_lower: 'island',
-    full_name: 'Island [unh-137]',
-    artist: 'John Avon',
-    scryfall_uri: 'https://scryfall.com/card/unh/137/island?utm_source=api',
-    rarity: 'common',
-    oracle_text: '({T}: Add {U}.)',
-    _id: '0c4eaecf-dd4c-45ab-9b50-2abe987d35d4',
-    cmc: 0,
-    legalities: {
-      Legacy: true,
-      Modern: true,
-      Standard: true,
-      Pauper: true
-    },
-    parsed_cost: [''],
-    colors: [],
-    type: 'Basic Land — Island',
-    tcgplayer_id: 37875,
-    image_small: 'https://img.scryfall.com/cards/small/front/0/c/0c4eaecf-dd4c-45ab-9b50-2abe987d35d4.jpg?1565989364',
-    image_normal: 'https://img.scryfall.com/cards/normal/front/0/c/0c4eaecf-dd4c-45ab-9b50-2abe987d35d4.jpg?1565989364',
-    art_crop: 'https://img.scryfall.com/cards/art_crop/front/0/c/0c4eaecf-dd4c-45ab-9b50-2abe987d35d4.jpg?1565989364',
-    colorcategory: 'l',
-  }
-};
 
 beforeEach(() => {
   sinon.stub(Cube, "findOne");
@@ -301,8 +144,8 @@ test("getBasics returns the expected set of basic lands", () => {
   var expected = {};
   var exampleLand, expectedLandObject;
   basicLands.forEach(function(name) {
-    mockCarddict[mockNameToId[name.toLowerCase()]] = exampleBasics[name.toLowerCase()];
-    exampleLand = exampleBasics[name.toLowerCase()];
+    mockCarddict[mockNameToId[name.toLowerCase()]] = landsfixture.exampleBasics[name.toLowerCase()];
+    exampleLand = landsfixture.exampleBasics[name.toLowerCase()];
     expectedLandObject = {
       // copy necessary because getBasics modifies carddb (bad)
       'details': JSON.parse(JSON.stringify(exampleLand))
